+++ cmaqsatproc/config.py
@@ -1,0 +1,5 @@
+__init__ = ['rcParams']
+
+rcParams = {
+    'satellite.cmr_to_level3.method': 'download',
+}